--- Calculating Interest.py
@@ -1,12 +1,0 @@
-def simpleInterest():
-  principal = float(input("Enter the principal amount: "))
-  rate = float(input("Enter the rate per year: "))
-  no_of_years = float(input("Enter the number of years: "))
-  SI = ( principal * no_of_years * rate ) / 100
-  print("Simple interest is:", SI)
-
-print("What would you like to calculate?")
-print("1. Simple Interest")
-Input = int(input("Enter the corresponding number: "))
-if Input == 1:
-  simpleInterest()
