+++ Programs/Stone-paper-scissor.py
