+++ Programs/checkPalindrome.py
